Lint, codestyle
