+++ server/sdk/types/ITeamService.ts
@@ -61,5 +61,5 @@ export interface ITeamService {
 	getOneById(teamId: string): Promise<ITeam | undefined>;
 	getOneByName(teamName: string): Promise<ITeam | null>;
 	getMatchingTeamRooms(teamId: string, rids: Array<string>): Promise<Array<string>>;
-	autocomplete(uid:string, name: string): Promise<Array<IRoom>>;
+	autocomplete(uid: string, name: string): Promise<Array<IRoom>>;
 }
